Kinde Auth Migration

+++ actions/admin/qrcode/delete.ts
@@ -2,9 +2,20 @@
 
 import { ActionReturnType } from "@/interfaces/actionReturnType";
 import { prisma } from "@/lib/db";
+import { getKindeServerSession } from "@kinde-oss/kinde-auth-nextjs/server";
 import { Prisma } from "@prisma/client";
 
 export async function DeleteQRCode(id: string): Promise<ActionReturnType> {
+  const { getUser, getRoles } = getKindeServerSession();
+  const user = await getUser();
+  const roles = await getRoles();
+  const isAdmin = roles?.some((role) => role.key === "admin") || false;
+  if (!isAdmin || !user) {
+    return {
+      message: "Not authorized.",
+      status: 401,
+    };
+  }
   try {
     await prisma.qrCode.delete({
       where: {
